requirements.txt
langchain
langsmith
langchain-openai
notebook
pinecone-client
python-dotenv
requests
tqdm
writerai
streamlit